jwttoken send & receive success

+++ controllers/api/v1/posts_api.js
@@ -31,20 +31,29 @@ module.exports.index=function(req,res){
 module.exports.destroy = function(req, res){
     Post.findById(req.params.id).then(
         (post)=>{
-           
-               Post.deleteOne({_id:post._id}).then(()=>{
+           if(post.user==req.user.id){
+            Post.deleteOne({_id:post._id}).then(()=>{
                 Comment.deleteMany({post: req.params.id}).then();
 
                })
                console.log("delete kru main post");
                return res.json(200,{
-                message:"posts nd comments deleted successfuly"
+                message:"Your posts & comments deleted successfully. Have a good day !:)"
                });
+            
+           }
+           else{
+            console.log("Unauthorized");
+            return res.json(200,{
+             message:"Dont delete someone else's post. Thank you :)"
+            });
+           }
+              
                 
                     
                 }).catch(err=>{
                 return res.json(500,{
-                    message:'internal servere error'
+                    message:'OMG ! Internal servere error'
                 });
             })
